seclopzbot/requirements.txt
click==6.7
flask==1.0
slackclient==2.1
